change from Float to Int

+++ server/graphql/schema/chi-tiet-phieu-nhap.schema.js
@@ -9,7 +9,7 @@ const typeDefs = gql`
         id: ID!
         maPhieuNhap: String!
         sanPham: SanPham
-        soLuongNhap: Float!
+        soLuongNhap: Int!
         donGiaNhap: Int!
         ngayNhap: Date!
         ghiChu: String
